# docs/api/java-gateway.openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/stealthguard/telemetry":{"post":{"tags":["telemetry-controller"],"operationId":"telemetry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelemetryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DecisionResponse"}}}}}}},"/stealthguard/session/init":{"post":{"tags":["session-controller"],"operationId":"init","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionInitResponse"}}}}}}},"/stealthguard/challenge/{sessionId}/respond":{"post":{"tags":["telemetry-controller"],"operationId":"respond","parameters":[{"name":"sessionId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChallengeResponseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DecisionResponse"}}}}}}},"/stealthguard/decision/{sessionId}":{"get":{"tags":["telemetry-controller"],"operationId":"decision","parameters":[{"name":"sessionId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DecisionResponse"}}}}}}},"/stealthguard/admin/stats":{"get":{"tags":["admin-controller"],"operationId":"stats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}}}}},"/stealthguard/admin/sessions":{"get":{"tags":["admin-controller"],"operationId":"sessions","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageSession"}}}}}}}},"components":{"schemas":{"KeystrokeDto":{"type":"object","properties":{"key":{"type":"string"},"downTime":{"type":"number","format":"double"},"upTime":{"type":"number","format":"double"}}},"MetaDto":{"type":"object","properties":{"userAgent":{"type":"string"},"viewportWidth":{"type":"integer","format":"int32"},"viewportHeight":{"type":"integer","format":"int32"},"timezoneOffset":{"type":"integer","format":"int32"},"inputModality":{"type":"string"}}},"MouseMoveDto":{"type":"object","properties":{"x":{"type":"number","format":"double"},"y":{"type":"number","format":"double"},"t":{"type":"number","format":"double"}}},"TelemetryRequest":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"},"page":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"sdkVersion":{"type":"string"},"privacyMode":{"type":"string"},"keystrokes":{"type":"array","items":{"$ref":"#/components/schemas/KeystrokeDto"},"maxItems":5000,"minItems":0},"mouseMoves":{"type":"array","items":{"$ref":"#/components/schemas/MouseMoveDto"},"maxItems":5000,"minItems":0},"touchMoves":{"type":"array","items":{"$ref":"#/components/schemas/MouseMoveDto"},"maxItems":5000,"minItems":0},"clicks":{"type":"array","items":{"$ref":"#/components/schemas/MouseMoveDto"},"maxItems":5000,"minItems":0},"features":{"type":"object","additionalProperties":{"type":"number","format":"double"}},"meta":{"$ref":"#/components/schemas/MetaDto"}},"required":["sessionId"]},"DecisionResponse":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"},"decision":{"type":"string"},"humannessScore":{"type":"number","format":"double"},"modelVersion":{"type":"string"},"reasonCodes":{"type":"array","items":{"$ref":"#/components/schemas/ReasonCodeDto"}}}},"ReasonCodeDto":{"type":"object","properties":{"code":{"type":"string"},"weight":{"type":"number","format":"double"}}},"SessionInitResponse":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"},"issuedAt":{"type":"string","format":"date-time"}}},"ChallengeResponseRequest":{"type":"object","properties":{"challengeType":{"type":"string"},"response":{"type":"string"}}},"PageSession":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Session"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"}}},"Session":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"page":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"userAgent":{"type":"string"},"viewportWidth":{"type":"integer","format":"int32"},"viewportHeight":{"type":"integer","format":"int32"},"timezoneOffset":{"type":"integer","format":"int32"},"inputModality":{"type":"string"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"empty":{"type":"boolean"},"unsorted":{"type":"boolean"}}}}}}