# java-gateway/src/main/java/org/stealthguard/gateway/dto/SessionInitResponse.java
package org.stealthguard.gateway.dto;

import java.time.Instant;
import java.util.UUID;

/** Session-init response (SPEC §6.1). */
public record SessionInitResponse(UUID sessionId, Instant issuedAt) {
}